SQOOP-39. (SIP-6) Session Management API. Added SessionStorage API. Added SessonData API. Added SessionStorageFactory. Added SessionTool for create/delete/execute/show/list operations on sessions. SqoopOptions can read and write all "sticky" state to a Properties instance. Added HsqldbSessionStorage to implement SessionStorage API. Added AutoHsqldbStorage to auto-instantiate a local metastore for the user. Added client metastore connection parameters to sqoop-site.xml. Added metastore tool (MetastoreTool). Added HsqldbMetaStore for standalone metastore instance. Added metastore properties to sqoop-default.xml. Added TestSessions unit tests of session API. Renamed conf/sqoop-default.xml to conf/sqoop-site-template.xml. Added conf/.gitignore for sqoop-site.xml.
Tests run:

Tested all metastore operations on an import session.
Tested that ~/.sqoop/-based storage will be auto-created by the metastore.
Tested that 'sqoop metastore'-based metastores can be connected to
by external clients.
Tested that 'sqoop metastore --shutdown' will gracefully shut down a running
metastore instance.
Tested that passwords are not stored in the metastore by default, and the
user is prompted for the password when executing that saved session.