SQOOP-187. Allow skipping end-of-record delimiter.
The SqoopRecord.toString() and SqoopRecord.toString(DelimiterSet) methods
always append an end-of-record delimiter. Sqoop uses its own OutputFormat
when rendering these to text files, so that the user's delimiters are
preserved.

Other users could use this OutputFormat when working with SqoopRecord
instances in their own MapReduce code, but it would also be nice to "play
nice" with TextOutputFormat in the event that the intent is
newline-terminated records.

This patch allows users to suppress end-of-record delimiter generation when
formatting records with toString.

(Aaron Kimball via Arvind Prabhakar)