CLOUDERA-BUILD. Enable Sqoop unit tests against MR1.
If -Dmr1.test=mr1 is given, the MR1 hadoop-core and hadoop-test jars
are put in the test classpath, ahead of the MR2 jars. While it's
possible to jury-rig the build to actually compile against MR1, this
ends up not actually providing much value and is a lot harder to
configure. This approach will allow for building against MR2 but
running the unit tests against MR1, which is the real goal.