JENKINS-657 Creating missing unittest for sqoop for all supprted branches
Currently sqoop unittest jobs are setting up environments under jenkins console (shell script).
And it also depends on "tool env" plugin from jenkins to get environments.
Idea is, cleanup stuffs and setup all environment outside jenkins so it does not look very clutter
and if we need to make any environment changes it would be jenkins independent.
Meaning you do not need to update jenkins job for env changes.

Also this script would be able to figure out environment on base of command line arguments and TOOLCHAIN_HOME passed as argument.
In older setup some of the settings used under sqoop unittest jobs are not really being used.
For example:  ANT_1_8_2_HOME which is not being used and same with JAVA_6_HOME etc..

We are going to use similar script for most of the components , Hadoop is already using it
http://github.mtv.cloudera.com/CDH/hadoop/blob/cdh5-2.6.0/cloudera/test-stable.sh
for easy transition towards newer cloudbees jenkins for all unittest jobs for all components.
Changes to the jobs will be done via source control rathern than UI.

Summarize benefits:
1) Jenkins independent changes for environment which will happen under sqoop repo (for example: cloudera/test-stable.sh)
2) Much cleaner and uniform way of running unittests across components.
3) Easier/programatic migration to newer cloudbees jenkins and creating unittest jobs from newer releases.
4) Environment being created using TOOLCHAIN_HOME

Change-Id: If5dc391b51d94680b8e381df21b073a2f1e253e1